Refactor inflight removal modules for #4201
